--- lib/lasso_web/controllers/hook_controller.ex
@@ -1,16 +1,0 @@
-defmodule LassoWeb.HookController do
-  use LassoWeb, :controller
-  require Logger
-
-  def request(conn, %{"uuid" => uuid}) do
-    request = Lasso.Request.from(conn)
-
-    with {:ok, _value} <- Lasso.Hook.get(uuid),
-         :ok <- Lasso.Hook.add(uuid, request) do
-      render(conn, "request.json", uuid: uuid)
-    else
-      {:error, :no_such_key, _} ->
-        resp(conn, 410, "No such lasso")
-    end
-  end
-end
