+++ lib/lasso_web/templates/page/index.html.eex
@@ -14,7 +14,7 @@
   <p>When you create a “lasso”, you’ll get a unique URL that you can use as the destination for HTTP requests. Lasso will receive and store HTTP requests to this endpoint, so you can view and debug the body/headers/params.</p>
   
   <p class="lead justify-content-center">
-    <%= form_for @conn, Routes.lasso_path(@conn, :new), [method: :post], fn _f -> %>
+    <%= form_for @conn, Routes.lasso_view_path(@conn, :new), [method: :post], fn _f -> %>
         <%= submit "Create new lasso", class: "btn btn-dark btn-lg" %>
     <% end %>
   </p>
