--- lib/lasso_web/views/hook_view.ex
@@ -1,7 +1,0 @@
-defmodule LassoWeb.HookView do
-  use LassoWeb, :view
-
-  def render("request.json", %{uuid: uuid}) do
-    %{data: %{id: uuid}}
-  end
-end
