+++ lib/lasso_web/views/lasso_view.ex
@@ -1,3 +1,7 @@
 defmodule LassoWeb.LassoView do
   use LassoWeb, :view
+
+  def render("request.json", %{uuid: uuid}) do
+    %{data: %{id: uuid}}
+  end
 end
