+++ lib/lasso_web/views/lasso_view_view.ex
@@ -1,0 +1,3 @@
+defmodule LassoWeb.LassoViewView do
+  use LassoWeb, :view
+end
