+++ test/lasso_web/controllers/lasso_controller_test.exs
@@ -1,9 +1,9 @@
-defmodule LassoWeb.HookControllerTest do
+defmodule LassoWeb.LassoControllerTest do
   use LassoWeb.ConnCase
 
-  describe "GET /hooks/:uuid" do
+  describe "GET /lasso/:uuid" do
     test "returns 410 for unknown", %{conn: conn} do
-      conn = get(conn, "/hooks/123")
+      conn = get(conn, "/lasso/123")
       assert response(conn, 410) =~ "lasso"
     end
   end
